refactor: move AI agent implementation to a dedicated request handler which is called from outbound connector
